nvim: Access VS Code snippets from nvim
By utilizing the LuaSnip extension for Neovim, the VS Code style
snippets are accessed. Placing the `package.json` file in the std path
of neovim config did not work for some reason. Thus, it is placed in the
path of Code config. Even though the Code files are synced over
SettingsSync, the files are added to dotfiles for convenience.