fix(project): add `init.lua` pattern to project
Since the `nvim/` directory is in `dotfiles` repo, based on the `.git`
folder the project plugin was reverting the root directory of `dotfiles`.
Due to being inside `.config/` hidden file, Neovim files were only be
accessible with hidden file flag of telescope. In order to prevent this
problem, `init.lua` pattern is added to `project.nvim` patterns.