fix(nvim): adjust catppuccin for ts
Catppuccin theme was having different styles for parameter names and
their counterparts in the functions. Even though they are the same
variable, their styles were different due to the limitations of
treesitter. Tree-sitter does not have LSP information, so semantic
highlight is not possible, currently even though there are PR who
addresses this issue in neovim. Therefore, `TSParameter` style is reset
to plain text style to match the `variable` highlight. As suggested in
the documents of catppuccin, `additional_regex_highlighting` is also
disabled.