fix(zathura): remove `best-fit` adjust open option
The adjust-open option to set the "best fit" to initial
launch was causing problems in adjusting afterwards. Hence,
it is commented out in the rc file.

In this process default recolor option is also commented out,
since by default I read them in the original form and barely
use the dark format.